requirements.txt
requests
pandas
seaborn
matplotlib
scipy
python-dotenv
networkx>=3.0
